==> 01_1_preprocessing.Rmd <==
---
title: "Preprocessing"
author: "Seraphina Shi"
output:
  html_document:
    toc: true
    toc_float: true
---

```{r setup, include = FALSE}
setwd("/Users/seraphinashi/Desktop/PAHS/Xuanwei Study coal and DNA metylation study/Seraphina's folder/PAHS-scripts")

plotFolder <- "../images/01_preprocessing/"
if(!file.exists(plotFolder)) dir.create(plotFolder,recursive=TRUE)

knitr::opts_chunk$set(
  cache=FALSE, autodep=FALSE, warning=FALSE, message=FALSE, echo=FALSE,
  results = 'markup', dev='png', dpi=150, fig.align = "center", fig.path=plotFolder,
  cache.path=".cache/",
  duplicate.label="allow"
)

```

```{r}
source('shared_commands.R')
```


# Load & Merge Data 
```{r}
# from clocks, fuels, toves, ambient exposures, and urinary measurements
dat <- read.csv("../data/0.calculate_clocks/horvathcalculator_beta_input.output.csv")
dat <- dat[dat$Gender != "M", ]
dat <- dat[!(grepl("CR" ,dat$SID) & dat$Group == "LEX"), ] # remove controls that are not clean

clusCUR6 <- read.csv("../data/new_exposure_files/LEX_clusCUR6.csv") %>% 
  select(- Visit) %>%
  unique()
clusCHLD5 <- read.csv("../data/new_exposure_files/LEX_clusCHLD5.csv") %>% 
  select(- Visit) %>%
  unique()
clusCUM6 <- read.csv("../data/new_exposure_files/LEX_clusCUM6.csv") %>% 
  select(- Visit) %>%
  unique()
clusMEAS6 <- read.csv("../data/new_exposure_files/LEX_clusMEAS6.csv") %>% 
  select(- Visit) %>%
  unique()
clusURI5 <- read.csv("../data/new_exposure_files/LEX_clusURI5.csv") 

dat <- dat %>% 
  merge(clusCUR6, by=c("SbjctD"), all.x = T) %>% 
  merge(clusCHLD5, by=c("SbjctD"), all.x = T) %>% 
  merge(clusCUM6, by=c("SbjctD"), all.x = T) %>% 
  merge(clusMEAS6, by=c("SbjctD"), all.x = T) %>% 
  merge(clusURI5, by=c("SbjctD", "Visit"), all.x = T) 
```

```{r}
names(dat)[which(names(dat)=="bap_conc_air"):which(names(dat)=="PYR_conc_")] <- ambient_exp_vars
names(dat)[which(names(dat)=="Benzo.a.anthracene...Chrysene..ng.g.Creatinine."):which(names(dat)=="Pyrene..ng.g.Creatinine.")] <- urinary_exp_vars
dat$Age = dat$age

dat <- dat %>% 
  select(unname(unlist(var_name_list))) %>% 
  filter((!is.na(DNAmAge)) | (!is.na(DNAmAgeHannum)) | (!is.na(DNAmPhenoAge)) | (!is.na(DNAmAgeSkinBloodClock)) | (!is.na(DNAmGrimAge)) | (!is.na(DNAmTL))) %>%
  unique()
```

```{r}
cat(sprintf("We have %s observations with %s columns: \n", dim(dat)[1], dim(dat)[2]))
print(var_name_list)
```




# Duplicated Observations
```{r}
SID_dup <- dat$SID[duplicated(dat$SID)]
dat_dup <- dat[dat$SID %in% SID_dup, ] %>% arrange(SID)
```
We have 3 subjects (SIDs: `r SID_dup`) with duplicated observations.

## Clocks stable?
In section, we check if the clocks are stable by comparing them in the duplicated observations. 


### Concordance between pairs
Evaluate correlation/concordance between the duplicate pairs 
```{r concordance_between_duplicated_pairs, fig.width=6, fig.height=4}
dat_dup_clocks <- dat_dup[, c("SID",clocks_vars)] 

dat_dup_tmp <- dat_dup_clocks %>% mutate(ob_n = rep(c(1,2),3) )
dat_dup_tmp <- dat_dup_tmp %>% tidyr::gather(clock, measurement, DNAmAge:DNAmTL)

dat_dup_tmp$clock[dat_dup_tmp$clock == clocks_vars[1]] = clock_names[1]
dat_dup_tmp$clock[dat_dup_tmp$clock == clocks_vars[2]] = clock_names[2]
dat_dup_tmp$clock[dat_dup_tmp$clock == clocks_vars[3]] = clock_names[3]
dat_dup_tmp$clock[dat_dup_tmp$clock == clocks_vars[4]] = clock_names[4]
dat_dup_tmp$clock[dat_dup_tmp$clock == clocks_vars[5]] = clock_names[5]
dat_dup_tmp$clock[dat_dup_tmp$clock == clocks_vars[6]] = clock_names[6]


dat_dup_tmp <- dat_dup_tmp[dat_dup_tmp$ob_n==1, ] %>%  
  mutate(measurement2 = dat_dup_tmp$measurement[dat_dup_tmp$ob_n==2])


ggplot(dat_dup_tmp) +
  geom_abline(slope=1, intercept = 0, lty=2, color="gray") +
  geom_point(aes(x=measurement, y=measurement2, color=clock)) +
  labs(title="Compare the clocks of the duplicated observations",
       x="One observation", y="Another observation") +
  scale_color_manual(values = colors_clock[1:6]) +
  theme_bw() +
  theme(plot.title = element_text(hjust = 0.5)) 

ggsave("concordance_between_duplicated_pairs.png")
```

```{r concordance_between_duplicated_pairs_1, fig.width=6, fig.height=4}
dat_dup_tmp <- dat_dup_tmp[dat_dup_tmp$clock != "DNAmTL",]
ggplot(dat_dup_tmp) +
  geom_abline(slope=1, intercept = 0, lty=2, color="gray") +
  geom_point(aes(x=measurement, y=measurement2, color=clock)) +
  xlim(min(dat_dup_tmp$measurement, dat_dup_tmp$measurement2), 
       max(dat_dup_tmp$measurement, dat_dup_tmp$measurement2)) +
  ylim(min(dat_dup_tmp$measurement, dat_dup_tmp$measurement2), 
       max(dat_dup_tmp$measurement, dat_dup_tmp$measurement2)) +
  labs(title="Compare the clocks of the duplicated observations",
       x="One observation", y="Another observation") +
  scale_color_manual(values = colors_clock[1:6]) +
  theme_bw() +
  theme(plot.title = element_text(hjust = 0.5))
ggsave("concordance_between_duplicated_pairs_1.png")
```


```{r concordance_between_duplicated_pairs_accel, fig.width=6, fig.height=4}
dat_dup_clocks_eaa <- dat_dup[, c("SID",EAAs_vars)] 

dat_dup_tmp <- dat_dup_clocks_eaa %>% mutate(ob_n = rep(c(1,2),3) )
dat_dup_tmp <- dat_dup_tmp %>% tidyr::gather(clock, measurement, AgeAccelerationResidual:EEAA)

dat_dup_tmp$clock[dat_dup_tmp$clock == EAAs_vars[1]] = EAAs_names[1]
dat_dup_tmp$clock[dat_dup_tmp$clock == EAAs_vars[2]] = EAAs_names[2]
dat_dup_tmp$clock[dat_dup_tmp$clock == EAAs_vars[3]] = EAAs_names[3]
dat_dup_tmp$clock[dat_dup_tmp$clock == EAAs_vars[4]] = EAAs_names[4]
dat_dup_tmp$clock[dat_dup_tmp$clock == EAAs_vars[5]] = EAAs_names[5]
dat_dup_tmp$clock[dat_dup_tmp$clock == EAAs_vars[6]] = EAAs_names[6]


dat_dup_tmp <- dat_dup_tmp[dat_dup_tmp$ob_n==1, ] %>%  
  mutate(measurement2 = dat_dup_tmp$measurement[dat_dup_tmp$ob_n==2])


ggplot(dat_dup_tmp) +
  geom_abline(slope=1, intercept = 0, lty=2, color="gray") +
  geom_point(aes(x=measurement, y=measurement2, color=clock)) +
  labs(title="Compare the age accelarations of clocks \n between the duplicated observations",
       x="One observation", y="Another observation") +
  scale_color_manual(values = colors_clock) +
  theme_bw() +
  theme(plot.title = element_text(hjust = 0.5)) 

ggsave("concordance_between_duplicated_pairs(acceleration).png")
```


### Absolute differences in two observations
Clocks / Age Accelerations
```{r}
diff_df <- matrix(c(as.numeric(dat_dup_clocks_eaa[1,-1]-dat_dup_clocks_eaa[2,-1]), 
                    as.numeric(dat_dup_clocks_eaa[3,-1]-dat_dup_clocks_eaa[4,-1]),
                    as.numeric(dat_dup_clocks_eaa[5,-1]-dat_dup_clocks_eaa[6,-1])),
                  nrow=3, byrow = T)
colnames(diff_df) <- EAAs_names
diff_df <- abs(diff_df)
diff_df

cat("\nSummary: :\n")
diff_df_s <- matrix(c(as.numeric(apply(diff_df, 2, mean)),
                     as.numeric(apply(diff_df, 2, sd))),
                     nrow=2, byrow=T)
colnames(diff_df_s) <- EAAs_names
row.names(diff_df_s) <- c("mean", "sd")

diff_df_s

```

## ICC of each clock
```{r}
library(psych)
dat_dup_clocks_eaa <- dat_dup_clocks_eaa %>% arrange(SID)
icc_clock_eaa <- function(clock){
  
  data_tmp <- data.frame(obs1 = dat_dup_clocks_eaa[c(1,3,5), clock],
                         obs2 = dat_dup_clocks_eaa[c(2,4,6), clock])
  return(ICC(data_tmp)$results)
}
for(i in 1:length(EAAs_vars)){
  cat("\n", EAAs_names[i], ": \n")
  results <- icc_clock_eaa(EAAs_vars[i])
  print(results[,1:6])
}
```


## Remove duplicated observations
```{r}
dat_tmp <- dat
dat <- dat[-which(duplicated(dat$SID)),]
```
Number of observations removed: `r nrow(dat_tmp) - nrow(dat)`


# Missing values
## For each exposure measurements:
Number of missing values in baseline covariates (confounders):
```{r}
sapply(dat[,unlist(var_name_list$confounders)], function(x) sum(is.na(x)))
```
Number of missing values in ambient exposure measurements covariates:
```{r}
sapply(dat[,c(ambient_exp_vars)], function(x) sum(is.na(x)))
```
Number of missing values in urinary measurements covariates:
```{r}
sapply(dat[,c(urinary_exp_vars)], function(x) sum(is.na(x)))
```

Number of missing values in clustering covariates:
```{r}
sapply(dat[,unlist(var_name_list$cluster_exp)], function(x) sum(is.na(x)))
```

## For each observations:
Observations with different number of total missing measurements:
```{r}
count_miss_per_obs <- function(df){
  tb <- t(as.matrix(table(rowSums(is.na(df)))))
  tb <- matrix(c(as.numeric(colnames(tb)), as.numeric(tb[1,])), nrow=2, byrow=T)
  rownames(tb) <- c("Number of missing measurements:", "Number of observations:")
  colnames(tb) <- rep("", ncol(tb))
  print(tb)
}
count_miss_per_obs(dat)
```

Observations with different number of missing air exposure measurements:
```{r}
count_miss_per_obs(dat[,ambient_exp_vars])
```

Observations with different number of missing urinary measurements:
```{r}
count_miss_per_obs(dat[,urinary_exp_vars])
```


# Control Group
## Table with summarize info 
```{r fig.width=6, fig.height=4}
dat[dat$Group != "LEX",] %>%  
  select(-SID) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

## Clock values
```{r clocks_vs_groups, fig.height=3, fig.width=8}
library(reshape2)
meltdf <- melt(data = dat %>% select(c("SID", "Group", "Age", clocks_vars[-6])), id.vars = c("SID", "Group"))
ggplot(meltdf, aes(x=as.factor(variable), y=value, fill=Group))+
  geom_boxplot(position = position_dodge(width = 1)) +  
  theme_bw() +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(x="Age and Clocks", title = "Compare ages and clocks between two groups")
ggsave("age_clocks_vs_groups.png", width=10, height=3, dpi=1200)
```

## Correlogram Between clocks
```{r correlogram_clocks_control, fig.width=10, fig.height=10}
library(GGally)
ggpairs(dat[dat$Group=="LCW Clean Ctrl",c("Age", clocks_vars)]) +
  labs(title = "Correlogram of all clocks \n Control Group") +
  theme(plot.title = element_text(hjust=0.5, size=24)) 
ggsave("correlogram_clocks_control.png", dpi=1200)
```

## Scatterplots of age vs each epigenetic age
```{r}
median_absolute_error <- function(clock_var){
  return( median(abs(dat[,clock_var] - dat$Age)))
}

MAE <- sapply(clocks_vars[1:5], median_absolute_error)

age_vs_clock <- function(df_group, clock, clock_name) {
  if(is.na(df_group)){
    df_tmp <- dat
  } else {
    df_tmp <- dat[dat$Group == df_group,]
  }
  cor_test <- cor.test(df_tmp$Age, df_tmp[,clock])
  r_pearson <- paste0("r=", round(cor_test$estimate, 2))
  p_val <- ifelse(cor_test$p.value<0.0001, "P<0.0001", 
                  ifelse(cor_test$p.value<0.001, "P<0.001", 
                         ifelse(cor_test$p.value<0.01, "P<0.01", 
                                paste0("P=", round(cor_test$p.value, 2)))))
  MAE <- paste0("MAE=", round(median_absolute_error(clock),2))
  ttl <- ifelse(clock == "DNAmTL", paste0("[", LETTERS[which(clocks_vars == clock)], "]: ",r_pearson, ", ", p_val),
                  paste0("[", LETTERS[which(clocks_vars == clock)], "]: ",r_pearson, ", ", p_val, ", ", MAE, " (years)"))
  pi <- ggplot(df_tmp, aes_string(x="Age", y=clock)) +
    geom_point() + 
    labs(y=clock_name, x="Chronological Age (years)",
         title = ttl) +
    geom_smooth(method = lm, se=T, col = "black") +
    theme(plot.title = element_text(size=10))
  return(pi)
}
```

```{r plots_clocks_vs_age_ctrol, fig.width=9, fig.height=6}
plots_clocks_vs_age <- mapply(age_vs_clock, rep("LCW Clean Ctrl", 6), clocks_vars, clock_names_extra, SIMPLIFY = FALSE)
do.call("grid.arrange", c(plots_clocks_vs_age, ncol=3))
```

# Remove the control group from data
```{r}
dat <- dat[dat$Group == "LEX", ]
```

# Output data
```{r}
# df <- dat[which(rowSums(is.na(dat))==0),]
write.csv(dat, "../data/data_cleaned.csv")
```
