==> 02_EDA.Rmd <==
---
title: "Exploration"
author: "Seraphina Shi"
output:
  html_document:
    toc: true
    toc_float: true
---

```{r setup, include = FALSE}
setwd("/Users/seraphinashi/Desktop/PAHS/Xuanwei Study coal and DNA metylation study/Seraphina's folder/PAHS-scripts")

plotFolder <- "../images/02_EDA/"
if(!file.exists(plotFolder)) dir.create(plotFolder,recursive=TRUE)

knitr::opts_chunk$set(
  cache=FALSE, autodep=FALSE, warning=FALSE, message=FALSE, echo=FALSE,
  results = 'markup', dev='png', dpi=150, fig.align = "center", fig.path=plotFolder,
  cache.path=".cache/",
  duplicate.label="allow"
)

```

```{r}
source('shared_commands.R')
```


# Load Data from the Clock Calculator
```{r}
df <- read.csv("../data/data_cleaned_final.csv") %>% select(-X)
```

```{r}
dup_sbjctD <- df$SbjctD[duplicated(df$SbjctD)]
df_sub1 = df[(df$SbjctD %in% dup_sbjctD & df$Visit == 1) | (! df$SbjctD %in% dup_sbjctD), ]
df_subs <- list("df_sub1" = df_sub1,
                "df_sub2" = df[(df$SbjctD %in% dup_sbjctD & df$Visit == 2) | (! df$SbjctD %in% dup_sbjctD),])
```


# 1. Table with summarize info [table 1]

## 1.1. Description of study population (all visits)
### Baseline characteristics (confounders)
```{r}
library(table1)
table1(~ Age + BMI + factor(county) + factor(ses)  + factor(edu), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```


### Epigenetic ages
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$clocks, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

### Epigenetic ages accelarations
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$EEAs, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

### Fuel/stove type exposures
```{r}
table1(as.formula(paste(" ~ ", paste(paste0("factor(", c(var_name_list$fuel_exp, var_name_list$stove_exp),")"), 
                                     collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

### 5MC exposures
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$PCH_5MC, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```


```{r}
print("Pearson pair-wise correlation:")
cor(df %>%  select(var_name_list$PCH_5MC) %>% na.omit(), 
    method = "pearson")
```
```{r}
print("Spearman pair-wise correlation:")
cor(df %>%  select(var_name_list$PCH_5MC) %>% na.omit(), 
    method = "spearman")
```

### Cluster-based exposures
#### clusCUR6
Clusters based on model-based exposure estimates at or shortly before the visit
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusCUR6, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

#### clusCHLD5
Clusters based on model-based exposure estimates accrued before age 18 
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusCHLD5, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

#### clusCUM6
Clusters based on model-based lifetime exposure estimates 
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusCUM6, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

#### clusMEAS6
Clusters based on pollutant measurements
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusMEAS6, collapse = " + "))), 
       data = df)
```

#### clusURI5
Clusters based on urinary biomarkers
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusURI5, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```



### Ambient exposures
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$ambient_exp, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

### Urinary biomarkers
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$urinary_exp, collapse = " + "))), 
       data = df,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```





## 1.2. Description of study population (first visit for all objects)
There are `r nrow(df)` visits with corresponding epigenetic ages available among `r length(unique(df$SbjctD))` female subjects. For these `r length(unique(df$SbjctD))` subjects, `r length(unique(df$SbjctD)) - length(dup_sbjctD)` have been visited once and `r length(dup_sbjctD)` have been visited twice. 

The following tables summarize all the information of the first visit of these `r length(unique(df$SbjctD))` subjects.

### Baseline characteristics (confounders)
```{r}
library(table1)
table1(~ Age + BMI + factor(county) + factor(ses)  + factor(edu), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```


### Epigenetic ages
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$clocks, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

### Epigenetic ages accelarations
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$EEAs, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

### Fuel/stove type exposures
```{r}
table1(as.formula(paste(" ~ ", paste(paste0("factor(", c(var_name_list$fuel_exp, var_name_list$stove_exp),")"), 
                                     collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

### 5MC exposures
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$PCH_5MC, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```


```{r}
print("Pearson pair-wise correlation:")
cor(df_sub1 %>%  select(var_name_list$PCH_5MC) %>% na.omit(), 
    method = "pearson")
```
```{r}
print("Spearman pair-wise correlation:")
cor(df_sub1 %>%  select(var_name_list$PCH_5MC) %>% na.omit(), 
    method = "spearman")
```

### Cluster-based exposures
#### clusCUR6
Clusters based on model-based exposure estimates at or shortly before the visit
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusCUR6, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

#### clusCHLD5
Clusters based on model-based exposure estimates accrued before age 18 
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusCHLD5, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

#### clusCUM6
Clusters based on model-based lifetime exposure estimates 
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusCUM6, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

#### clusMEAS6
Clusters based on pollutant measurements
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusMEAS6, collapse = " + "))), 
       data = df_sub1)
```

#### clusURI5
Clusters based on urinary biomarkers
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$cluster_exp$clusURI5, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```



### Ambient exposures
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$ambient_exp, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```

### Urinary biomarkers
```{r}
table1(as.formula(paste(" ~ ", paste(var_name_list$urinary_exp, collapse = " + "))), 
       data = df_sub1,
       render.continuous=c(.="Mean (SD)", 
                           .="Median (IQR)"))
```


## 1.3. EAA ~ confounders
```{r}
library(geepack)

preform_gee_exposure_prototype <- function(X, Y, adjust_cfd = T, log_trans_X=F) {
  tmp <- df
  for(i in 1:length(X)){
    X_i <- X[i]
    tmp <- tmp[!is.na(tmp[,X_i]),]
    if(log_trans_X) {
      tmp[,X_i] <- log(tmp[,X_i])
    }
  }
  
  X_form <- paste(X, collapse  = " + ")

  form <- ifelse(adjust_cfd, 
                 paste0(Y," ~ ", X_form, " + ", paste(var_name_list$confounders[-1], collapse  = " + ")),
                 paste0(Y,"~", X_form))
  form <- as.formula(form)
  
  print(paste("Fitting with", nrow(tmp), "observations."))
  geefit <- geeglm(formula = form, 
                data = tmp,
                id = SbjctD,
                family = gaussian,
                corstr = "exchangeable") # This says all pairs of responses within a subject are equally correlated


  coefs <- as.data.frame(summary(geefit)$coefficients)
  colnames(coefs) <- c("coefficient", "std", "t_val", "pval")

  coefs <- coefs %>%
    select(coefficient, std, pval) %>%
    mutate(ci_lower = coefficient - 1.96 * std,
           ci_upper = coefficient + 1.96 * std,
           sig_level = ifelse(pval <= 0.001, "<= 0.001",
                        ifelse(pval <= 0.01, "<= 0.01",
                               ifelse(pval <= 0.05, "<= 0.05", "> 0.05")))) %>% 
    select(coefficient, std, ci_lower, ci_upper, pval, sig_level)


  colnames(coefs) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  
  return(coefs)
}

```

```{r gee_EAA_confounders}
for(i in 1:length(var_name_list$EEAs)){
    cat("For", var_name_list$EEAs[i], ":")
    rlts = preform_gee_exposure_prototype(X = var_name_list$confounders,
                                          Y = var_name_list$EEAs[i],
                                          adjust_cfd = T)
    rlts$coefficient = round(as.numeric(rlts$coefficient), 4)
    rlts$std = round(as.numeric(rlts$std), 4)
    rlts$ci_lower = round(as.numeric(rlts$ci_lower), 4)
    rlts$ci_upper = round(as.numeric(rlts$ci_upper), 4)
    rlts$p_val = round(as.numeric(rlts$p_val), 4)
    print(rlts)
}
```


## 1.4. 5MC ~ confounders
```{r gee_5MC_confounders}
for(i in 1:length(var_name_list$PCH_5MC)){
    cat("For", var_name_list$PCH_5MC[i], ":")
    rlts = preform_gee_exposure_prototype(X = var_name_list$confounders,
                                          Y = var_name_list$PCH_5MC[i],
                                          adjust_cfd = T)
    rlts$coefficient = round(as.numeric(rlts$coefficient), 4)
    rlts$std = round(as.numeric(rlts$std), 4)
    rlts$ci_lower = round(as.numeric(rlts$ci_lower), 4)
    rlts$ci_upper = round(as.numeric(rlts$ci_upper), 4)
    rlts$p_val = round(as.numeric(rlts$p_val), 4)

    print(rlts)
}
```
## 1.5. clusCUR6 ~ confounders
```{r gee_clusCUR6_confounders}
for(i in 1:length(var_name_list$cluster_exp$clusCUR6)){
    cat("For", var_name_list$cluster_exp$clusCUR6[i], ":")
    rlts = preform_gee_exposure_prototype(X = var_name_list$confounders,
                                          Y = var_name_list$cluster_exp$clusCUR6[i],
                                          adjust_cfd = T)
    rlts$coefficient = round(as.numeric(rlts$coefficient), 4)
    rlts$std = round(as.numeric(rlts$std), 4)
    rlts$ci_lower = round(as.numeric(rlts$ci_lower), 4)
    rlts$ci_upper = round(as.numeric(rlts$ci_upper), 4)
    rlts$p_val = round(as.numeric(rlts$p_val), 4)

    print(rlts)
}
```

 
## Fuel Types Summary table
Current Fuel vs Birth Fuel + Cumulative
```{r}
library(reshape2)
dcast(df, brthFuel+cumFuel ~ curFuel)
```





# 2. PCA 
```{r pca_clocks, fig.height=10, fig.width=7}
library(ggfortify) 
library(tidyr)

df_complete <- df %>% select("Age", clocks_vars, ambient_exp_vars, urinary_exp_vars) %>% drop_na()

plot_pca_label <- function(clock, clock_name){
  pca <- prcomp(df_complete[,c(ambient_exp_vars, urinary_exp_vars)])
  pca_data <- as.data.frame(pca$x[,c(1,2)])
  pca_data$clock <- clock

  library("ggpubr")
  p_pca <- ggscatter(pca_data, x="PC1", y="PC2",
                    color = "clock", size = 1.5, alpha=0.5, legend = "right", ggtheme = theme_bw(),
                    xlab = paste0("Dim 1 (", summary(pca)$importance[2,1], "% )" ),
                    ylab = paste0("Dim 2 (", summary(pca)$importance[2,2], "% )" )) +
    labs(title = clock_name) +
    theme(plot.title = element_text(hjust=0.5))

  return(p_pca)
}

pca_plots <- vector(mode = "list", length = 6)
for(i in 1:6){
  if(i == 6){
    pca_plots[[i]] <- plot_pca_label(df_complete[,clocks_vars[i]], clock_names[i]) +
      scale_color_gradient(low="goldenrod1", high="steelblue4")
  } else {
    pca_plots[[i]] <- plot_pca_label(df_complete[,clocks_vars[i]], paste(clock_names[i], "Clock")) +
      scale_color_gradient(limits = c(20,90),
                           low="goldenrod1", high="steelblue4")
  }

}
grid.arrange(pca_plots[[1]], pca_plots[[2]], pca_plots[[3]], pca_plots[[4]], pca_plots[[5]], pca_plots[[6]],
             ncol=2, nrow=3,
             top = textGrob("Principal Component Analysis \n Colored by epigenetic ages",
                            gp=gpar(fontsize=20,font=3)))
ggsave("pca_clocks.png", dpi=1200)

```

```{r pca_age, fig.height=4, fig.width=5}
pca_plots_age <- plot_pca_label(df_complete$Age, "AGE") +
  labs(title = "Principal Component Analysis \n Colored by ages")
pca_plots_age
ggsave("pca_plots_age.png", dpi=1200)
```

# 3. Explore clocks
## 3.1. Clock values
```{r clocks_vs_groups, fig.height=3, fig.width=8}
meltdf <- melt(data = df %>% select(c("SID", "Group", "Age", clocks_vars[-6])), id.vars = c("SID", "Group"))
ggplot(meltdf, aes(x=as.factor(variable), y=value, fill=Group))+
  geom_boxplot(position = position_dodge(width = 1)) +  
  theme_bw() +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(x="Age and Clocks", title = "Compare ages and clocks between two groups")
ggsave("age_clocks_vs_groups.png", width=10, height=3, dpi=1200)
```


## 3.2. Correlogram
### Between clocks
All visits included
```{r correlogram_clocks, fig.width=10, fig.height=10}
library(GGally)
ggpairs(df[,c("Age", clocks_vars)]) +
  labs(title = "Correlogram of all epigenetic clocks") +
  theme(plot.title = element_text(hjust=0.5, size=24)) 
ggsave("correlogram_clocks.png", dpi=1200)
```



## 3.3. Median absolute error
```{r}
median_absolute_error <- function(clock_var){
  return( median(abs(df[,clock_var] - df$Age)))
}

MAE <- sapply(clocks_vars[1:5], median_absolute_error)
MAE
```


## 3.4. Scatterplots of age vs each epigenetic age
```{r plots_clocks_vs_age, fig.width=9, fig.height=6}
age_vs_clock <- function(df_group, clock, clock_name) {
  if(is.na(df_group)){
    df_tmp <- df
  } else {
    df_tmp <- df[df$Group == df_group,]
  }
  cor_test <- cor.test(df_tmp$Age, df_tmp[,clock])
  r_pearson <- paste0("r=", round(cor_test$estimate, 2))
  p_val <- ifelse(cor_test$p.value<0.0001, "P<0.0001", 
                  ifelse(cor_test$p.value<0.001, "P<0.001", 
                         ifelse(cor_test$p.value<0.01, "P<0.01", 
                                paste0("P=", round(cor_test$p.value, 2)))))
  MAE <- paste0("MAE=", round(median_absolute_error(clock),2))
  ttl <- ifelse(clock == "DNAmTL", paste0("[", LETTERS[which(clocks_vars == clock)], "]: ",r_pearson, ", ", p_val),
                  paste0("[", LETTERS[which(clocks_vars == clock)], "]: ",r_pearson, ", ", p_val, ", ", MAE, " (years)"))
  pi <- ggplot(df_tmp, aes_string(x="Age", y=clock)) +
    geom_point() + 
    labs(y=clock_name, x="Chronological Age (years)",
         title = ttl) +
    geom_smooth(method = lm, se=T, col = "black") +
    theme(plot.title = element_text(size=10))
  return(pi)
}

plots_clocks_vs_age <- mapply(age_vs_clock, rep(NA, 6), clocks_vars, clock_names_extra, SIMPLIFY = FALSE)
do.call("grid.arrange", c(plots_clocks_vs_age, ncol=3))
```



# 4. Explore Exposure Measurements
## 4.1. Correlogram

### 4.1.1. Between Ambient Exposure Measurements
```{r correlogram_ambient, fig.width=12, fig.height=12}
ggpairs(df[,c(ambient_exp_vars)]) +
  labs(title = "Correlogram of all Ambient Exposure Measurements") +
  theme(plot.title = element_text(hjust=0.5, size=24))
ggsave("correlogram_ambient.png", dpi=1200)
```

```{r correlation_ambient, fig.height=8, fig.width=8}
library(corrplot)
corrplot(cor( na.omit(df[,c(ambient_exp_vars)]) ),  addCoef.col = 'black', type = 'lower')
```

### 4.1.2. Between Urinary Exposure Measurements
```{r correlogram_urinary, fig.width=10, fig.height=10}
library(GGally)
ggpairs(df[,c(urinary_exp_vars)]) +
  labs(title = "Correlogram of all Urinary Exposure Measurements") +
  theme(plot.title = element_text(hjust=0.5, size=24))
ggsave("correlogram_urinary.png", dpi=1200)
```

```{r correlation_urinary, fig.height=6, fig.width=6}
corrplot(cor( na.omit(df[,c(urinary_exp_vars)]) ),  addCoef.col = 'black', type = 'lower')
```

### 4.1.3. Between Ambient and Urinary Exposure Measurements
```{r correlogram_ambient_urinary, fig.width=20, fig.height=20}
ggpairs(df[,c(ambient_exp_vars, urinary_exp_vars)]) +
  labs(title = "Correlogram of all Ambient and Urinary Exposure Measurements") +
  theme(plot.title = element_text(hjust=0.5, size=24))
# ggsave("correlogram_ambient_urinary.png", dpi=1200)
```

```{r correlation_ambient_urinary, fig.height=12, fig.width=12}
corrplot(cor( na.omit(df[,c(ambient_exp_vars, urinary_exp_vars)]) ),  addCoef.col = 'black', type = 'lower')
```

### 4.1.4. Between Clustered Exposure Measurements
```{r correlogram_clusteredExp, fig.width=20, fig.height=20}
ggpairs(df[,unname(unlist(var_name_list$cluster_exp))]) +
  labs(title = "Correlogram of all Clustered Exposure Measurements") +
  theme(plot.title = element_text(hjust=0.5, size=24))
```
```{r correlation_clusteredExp, fig.height=12, fig.width=12}
corrplot(cor( na.omit(df[,unname(unlist(var_name_list$cluster_exp))]) ),  addCoef.col = 'black', type = 'lower')
```


# 5. Clocks vs Measurements
## 5.1. Pairplots
### 5.1.1. Urinary Exposure Measurements
```{r clock_urine_exps, fig.width=12, fig.height=8}
df[,c("SID", clocks_vars, urinary_exp_vars)] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(ambient_exp, measurements, Benzanthracene_Chrysene_urine:Pyrene_urine) %>% 
  mutate(measurements = log(measurements),
         ambient_exp = str_replace(ambient_exp, '_urine', ''),
         clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() + 
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ ambient_exp, scales="free") +
  labs(title="Epigenetic Age vs (log transformed) Urinary Measurements", 
       y = "Epigenetic Age",
       x = "Urinary Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=24))
ggsave("clock_urine_exps.png", dpi=1200)
```

### 5.1.2. Ambient Exposure Measurements
```{r clock_air_exps, fig.width=16, fig.height=6}
df[,c("SID", clocks_vars, ambient_exp_vars)] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(ambient_exp, measurements, bap_air:PYR_air) %>%
  mutate(measurements = log(measurements),
         ambient_exp = str_replace(ambient_exp, '_air', ''),
         clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() +
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ ambient_exp, scales="free") +
  labs(title="Epigenetic Age vs (log transformed) Ambient Exposure", 
       y = "Epigenetic Age",
       x = "Ambient Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=24))
ggsave("clock_air_exps.png", dpi=1200)
```


```{r clock_air_exps_1, fig.width=12, fig.height=8}
df[,c("SID", clocks_vars, ambient_exp_vars[1:7])] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(ambient_exp, measurements, bap_air:BkF_air) %>%
  mutate(measurements = log(measurements),
         ambient_exp = str_replace(ambient_exp, '_air', ''),
         clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() +
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ ambient_exp, scales="free") +
  labs(title="Epigenetic Age vs (log transformed) Ambient Exposure", 
       y = "Epigenetic Age",
       x = "Ambient Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=24))
ggsave("clock_air_exps_1.png", dpi=1200)
```


```{r clock_air_exps_2, fig.width=12, fig.height=8}
df[,c("SID", clocks_vars, ambient_exp_vars[8:15])] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(ambient_exp, measurements, CHR_air:PYR_air) %>%
  mutate(measurements = log(measurements),
         ambient_exp = str_replace(ambient_exp, '_air', ''),
         clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() +
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ ambient_exp, scales="free") +
  labs(title="Epigenetic Age vs (log transformed) Ambient Exposure", 
       y = "Epigenetic Age",
       x = "Ambient Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=24))
ggsave("clock_air_exps_2.png", dpi=1200)
```

### 5.1.3. Clustered Exposure Measurements
```{r clock_cluster_exps, fig.width=20, fig.height=8}
df[,c("SID", clocks_vars, unname(unlist(var_name_list$cluster_exp)))] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(cluster_exp, measurements, CUR6_BC_PAH6:URI5_CHR) %>% 
  mutate(clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() + 
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ cluster_exp, scales="free") +
  labs(title="Epigenetic Age vs Clustered Exposure Measurements", 
       y = "Epigenetic Age",
       x = "Clustered Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=24))
ggsave("clock_cluster_exps.png", dpi=1200)
```

```{r clock_cluster_exps_clusCUR6, fig.width=8, fig.height=8}
df[,c("SID", clocks_vars, unname(unlist(var_name_list$cluster_exp$clusCUR6)))] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(cluster_exp, measurements, CUR6_BC_PAH6:CUR6_SO2) %>% 
  mutate(clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() + 
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ cluster_exp, scales="free") +
  labs(title="Epigenetic Age vs Clustered Exposure Measurements from clusCUR6", 
       y = "Epigenetic Age",
       x = "Clustered Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=14))
ggsave("clock_cluster_exps_clusCUR6.png", dpi=1200)
```

```{r clock_cluster_exps_clusCHLD5, fig.width=8, fig.height=8}
df[,c("SID", clocks_vars, unname(unlist(var_name_list$cluster_exp$clusCHLD5)))] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(cluster_exp, measurements, CHLD5_X7:CHLD5_SO2) %>% 
  mutate(clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() + 
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ cluster_exp, scales="free") +
  labs(title="Epigenetic Age vs Clustered Exposure Measurements from clusCHLD5", 
       y = "Epigenetic Age",
       x = "Clustered Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=14))
ggsave("clock_cluster_exps_clusCHLD5.png", dpi=1200)
```

```{r clock_cluster_exps_clusCUM6, fig.width=8, fig.height=8}
df[,c("SID", clocks_vars, unname(unlist(var_name_list$cluster_exp$clusCUM6)))] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(cluster_exp, measurements, CUM6_BC_NO2_PM:CUM6_SO2) %>% 
  mutate(clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() + 
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ cluster_exp, scales="free") +
  labs(title="Epigenetic Age vs Clustered Exposure Measurements from clusCUM6", 
       y = "Epigenetic Age",
       x = "Clustered Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=14))
ggsave("clock_cluster_exps_clusCUM6.png", dpi=1200)
```

```{r clock_cluster_exps_clusMEAS6, fig.width=8, fig.height=8}
df[,c("SID", clocks_vars, unname(unlist(var_name_list$cluster_exp$clusMEAS6)))] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(cluster_exp, measurements, MEAS6_BC_PM_RET:MEAS6_NO2_SO2) %>% 
  mutate(clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() + 
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ cluster_exp, scales="free") +
  labs(title="Epigenetic Age vs Clustered Exposure Measurements from clusMEAS6", 
       y = "Epigenetic Age",
       x = "Clustered Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=14))
ggsave("clock_cluster_exps_clusMEAS6.png", dpi=1200)
```

```{r clock_cluster_exps_clusURI5, fig.width=8, fig.height=8}
df[,c("SID", clocks_vars, unname(unlist(var_name_list$cluster_exp$clusURI5)))] %>% 
  tidyr::gather(clocks_name, clock_age, DNAmAge:DNAmTL) %>%
  tidyr::gather(cluster_exp, measurements, URI5_NAP_1M_2M:URI5_CHR) %>% 
  mutate(clocks_name = str_replace(clocks_name, 'clock', '')) %>% 
  ggplot(aes(y=clock_age, x=measurements)) +
  geom_point() + 
  geom_smooth(se = FALSE, method = "gam", formula = y ~ x) +
  facet_grid(clocks_name ~ cluster_exp, scales="free") +
  labs(title="Epigenetic Age vs Clustered Exposure Measurements from clusURI5", 
       y = "Epigenetic Age",
       x = "Clustered Exposure Measurements") +
  theme(plot.title=element_text(hjust=0.5, size=14))
ggsave("clock_cluster_exps_clusURI5.png", dpi=1200)
```
