==> 03_analysis.Rmd <==
---
title: "PASH Analysis"
author: "Seraphina Shi"
output:
  html_document:
    toc: true
    toc_float: true
---

```{r setup, include = FALSE}
setwd("/Users/seraphinashi/Desktop/Projects/PAHS/Xuanwei Study coal and DNA metylation study/Seraphina's folder/PAHS-scripts")

plotFolder <- "../images/03_analysis/"
if(!file.exists(plotFolder)) dir.create(plotFolder,recursive=TRUE)

knitr::opts_chunk$set(
  cache=FALSE, autodep=FALSE, warning=FALSE, message=FALSE, echo=FALSE,
  results = 'markup', dev='png', dpi=150, fig.align = "center", fig.path=plotFolder,
  cache.path=".cache/",
  duplicate.label="allow"
)

```

```{r}
source('shared_commands.R')
```

```{r}
df <- read.csv("../data/data_cleaned.csv") %>% select(-X)
df$SbjctD[is.na(df$SbjctD)] <- substr(df$SID[is.na(df$SbjctD)], 5, 7)
df$Visit[is.na(df$Visit)] <- substr(df$SID[is.na(df$Visit)], 4, 4)
```

```{r get_5MC_data}
library(readxl)
LEX_5MC_all <- read_excel("../data/new_exposure_files/LEX_5MC_all.xlsx")
# View(LEX_5MC_all)
LEX_5MC_all <- LEX_5MC_all %>% rename("SbjctD" = "subject_id", 
                                      "Visit" = "visit",
                                      "cur_5mc" = "sum.X5.methylchrysene_conc_final.imputed", 
                                      "cum_5mc" = "cumlife_5mc",
                                      "bir_5mc" = "birthyr_5mc",
                                      "cur_5mc_measured" = "_5_methylchrysene_conc_measured")


df <- merge(df, LEX_5MC_all, by = c("SbjctD", "Visit"), all.x=T)

var_name_list$PCH_5MC <- c("cur_5mc", "cum_5mc", "bir_5mc", "cur_5mc_measured")

dup_sbjctD <- df$SbjctD[duplicated(df$SbjctD)]
df_sub1 = df[(df$SbjctD %in% dup_sbjctD & df$Visit == 1) | (! df$SbjctD %in% dup_sbjctD), ]
df_subs <- list("df_sub1" = df_sub1,
                "df_sub2" = df[(df$SbjctD %in% dup_sbjctD & df$Visit == 2) | (! df$SbjctD %in% dup_sbjctD),])
```

```{r define_functions_1}
library(lmtest)
preform_lmtest_mix <- function(Y, X, log_trans_X = F, dat = df) {
  tmp <- dat
  for(i in 1:length(X)){
    X_i <- X[i]
    tmp <- tmp[!is.na(tmp[,X_i]),]
    if(log_trans_X) {
      tmp[,X_i] <- log(tmp[,X_i])
    }
  }
  
  X_form <- paste(X, collapse  = " + ")
  
  form_full <- as.formula(paste0(Y," ~ ", X_form, " + ", paste(var_name_list$confounders[-1], collapse  = " + ")))
  lmfit_full <- lm(form_full, data = tmp)
  form_reduced <- as.formula(paste0(Y," ~ ", paste(var_name_list$confounders[-1], collapse  = " + ")))
  lmfit_reduced <- lm(form_reduced, data = tmp)
  
  lrtest <- lrtest(lmfit_full, lmfit_reduced)

  return(round(lrtest$`Pr(>Chisq)`, 4)[2])
}

preform_lmtest_simple <- function(Y, X, dat = df, log_trans_X = F) {
  tmp <- dat
  for(i in 1:length(X)){
    X_i <- X[i]
    tmp <- tmp[!is.na(tmp[,X_i]),]
    if(log_trans_X) {
      tmp[,X_i] <- log(tmp[,X_i])
    }
  }
  
  X_form <- paste(X, collapse  = " + ")
  
  form_full <- as.formula(paste0(Y," ~ ", X_form))
  lmfit_full <- lm(form_full, data = tmp)
  form_reduced <- as.formula(paste0(Y," ~ 1"))
  lmfit_reduced <- lm(form_reduced, data = tmp)
  
  lrtest <- lrtest(lmfit_full, lmfit_reduced)

  return(round(lrtest$`Pr(>Chisq)`, 4)[2])
}


preform_regression <- function(Y, X, log_trans_X = F, adjust_cfd = F, dat = df) {
  tmp <- dat[!is.na(dat[,X]),]
  if(log_trans_X) {
    tmp[,X] <- log(tmp[,X])
  }
  
  form <- ifelse(adjust_cfd, 
                 paste0(Y," ~ ", X, " + ", paste(var_name_list$confounders[-1], collapse  = " + ")),
                 paste0(Y,"~",X))
  form <- as.formula(form)
  lmfit <- lm(form, data = tmp)
  coef <- summary(lmfit)$coefficients[2,1]
  coef_text <- as.character(round(coef,4))
  std <- summary(lmfit)$coefficients[2,2]
  std_text <- as.character(round(std,4))
  p_val <- summary(lmfit)$coefficients[2,4]
  sig_star <- ifelse(p_val<0.001, "***",
                ifelse(p_val<0.01, "** ",
                       ifelse(p_val<0.05, "*  ", "   ")))
  sig_level <- ifelse(p_val <= 0.001, "<= 0.001",
                ifelse(p_val <= 0.01, "<= 0.01",
                       ifelse(p_val <= 0.05, "<= 0.05", "> 0.05")))
  rlt <- paste0(coef_text, paste0(rep(" ", 7-nchar(coef_text)), collapse = ""),
                sig_star,
                "(", std_text, paste0(rep(" ", 6-nchar(std_text)), collapse = ""), ")")
  ci_l <- coef - 1.96 * std
  ci_u <- coef + 1.96 * std
  return(c(rlt, coef, ci_l, ci_u, p_val, sig_level))
}


preform_regression_exposure_prototype <- function(X, Y, log_trans_X=F) {
  
  tmp <- df
  for(i in 1:length(X)){
    X_i <- X[i]
    tmp <- tmp[!is.na(tmp[,X_i]),]
    if(log_trans_X) {
      tmp[,X_i] <- log(tmp[,X_i])
    }
  }
  
  X_form <- paste(X, collapse  = " + ")
  
  tmp <- tmp[tmp$Group == "LEX", ]

  form <- as.formula(paste0(Y," ~ ", X_form, " + ", paste(var_name_list$confounders[-1], collapse  = " + ")))
  lmfit <- lm(form, data = tmp)

  coefs <- as.data.frame(summary(lmfit)$coefficients)
  colnames(coefs) <- c("coefficient", "std", "t_val", "pval")
  coefs$pval_BHadj <- p.adjust(coefs$pval, method = "BH")

  coefs <- coefs %>%
    select(coefficient, std, pval, pval_BHadj) %>%
    mutate(ci_lower = coefficient - 1.96 * std,
           ci_upper = coefficient + 1.96 * std,
           sig_level = ifelse(pval <= 0.001, "<= 0.001",
                        ifelse(pval <= 0.01, "<= 0.01",
                               ifelse(pval <= 0.05, "<= 0.05", "> 0.05")))) %>% 
    select(coefficient, std, ci_lower, ci_upper, pval, sig_level, pval_BHadj)
  
  return(coefs)
}


sig_color <- c("<= 0.001" = "#7CAE00", "<= 0.01" = "#00BFC4", "<= 0.05" = "#C77CFF", "> 0.05" = "black")

make_forest_plot <- function(df_plot, exposure){
  df_plot$sig_level <- as.factor(df_plot$sig_level)
  p <- ggplot(data=df_plot, aes(x=EAAs, y=coefficient, ymin=ci_lower, ymax=ci_upper)) +
    geom_point(shape=24, size=2, aes(col = sig_level, fill = sig_level)) +
    geom_errorbar(shape=24, width=0.3, aes(col = sig_level)) +
    geom_hline(yintercept=0, lty=2,color="red") +  # add a dotted line at x=1 after flip
    # ylim(c(-2,2)) +  # not working
    coord_flip() +  # flip coordinates (puts labels on y axis)
    labs(x=expression(log[2]-PFAS))+
    labs(y="Change in year") +
    ggtitle(exposure)+
    theme(plot.title = element_text(size=8.6, hjust = 0.5),
          axis.title.y = element_blank(),
          panel.grid.major.x = element_blank(),
          axis.text.y = element_text(size=8),
          axis.title.x = element_text(size=8),
          axis.text = element_text(size=7)) +
    scale_color_manual(values = sig_color) +
    scale_fill_manual(values = sig_color)

  return(p)
}

make_forest_plot_lg <- function(df, exposure){
  df$sig_level <- as.factor(df$sig_level)
  p <- ggplot(data=df, aes(x=EAAs, y=coefficient, ymin=ci_lower, ymax=ci_upper)) +
    geom_point(shape=24, size=2, aes(col = sig_level, fill = sig_level)) +
    geom_errorbar(shape=24, width=0.3, aes(col = sig_level)) +
    geom_hline(yintercept=0, lty=2,color="red") +  # add a dotted line at x=1 after flip
    # ylim(c(-2,2)) +  # not working
    coord_flip() +  # flip coordinates (puts labels on y axis)
    labs(x=expression(log[2]-PFAS))+
    labs(y="Change in year") +
    ggtitle(exposure)+
    theme(plot.title = element_text(size=15, hjust = 0.5),
          axis.title.y = element_blank(),
          panel.grid.major.x = element_blank(),
          axis.text.y = element_text(size=10),
          axis.title.x = element_text(size=10),
          axis.text = element_text(size=10)) +
    scale_color_manual(values = sig_color) +
    scale_fill_manual(values = sig_color)

  return(p)
}


options(scipen = 999)
```

```{r define_functions_2}
library(geepack)

preform_gee <- function(Y, X, log_trans_X = F, adjust_cfd = F, dat = df) {
  tmp <- dat[!is.na(dat[,X]),]
  tmp$curStove[is.na(tmp$curStove)] <- "NA"
  if(log_trans_X) {
    tmp[,X] <- log(tmp[,X])
  }
  
  form <- ifelse(adjust_cfd, 
                 paste0(Y," ~ ", X, " + ", paste(var_name_list$confounders[-1], collapse  = " + ")),
                 paste0(Y,"~",X))
  form <- as.formula(form)
  geefit <- geeglm(formula = form, 
                data = tmp,
                id = SbjctD,
                family = gaussian,
                corstr = "exchangeable") # This says all pairs of responses within a subject are equally correlated
  coef <- summary(geefit)$coefficients[2,1]
  coef_text <- as.character(round(coef,4))
  std <- summary(geefit)$coefficients[2,2]
  std_text <- as.character(round(std,4))
  p_val <- summary(geefit)$coefficients[2,4]
  sig_star <- ifelse(p_val<0.001, "***",
                ifelse(p_val<0.01, "** ",
                       ifelse(p_val<0.05, "*  ", "   ")))
  sig_level <- ifelse(p_val <= 0.001, "<= 0.001",
                ifelse(p_val <= 0.01, "<= 0.01",
                       ifelse(p_val <= 0.05, "<= 0.05", "> 0.05")))
  rlt <- paste0(coef_text, paste0(rep(" ", 7-nchar(coef_text)), collapse = ""),
                sig_star,
                "(", std_text, paste0(rep(" ", 6-nchar(std_text)), collapse = ""), ")")
  ci_l <- coef - 1.96 * std
  ci_u <- coef + 1.96 * std
  return(c(coef, std, ci_l, ci_u, p_val, sig_level))
}


perform_group_gee <- function(X, adjust_cfd = F){
  plot_df_list <- list()
  lm_coef_ambient_plot_list <- list()
  for (j in 1:length(X)) {
    exp <- X[j]
    plot_df <- data.frame(matrix(NA, ncol = 6))
    for(i in 1:length(var_name_list$EEAs)){
      rlts = preform_gee(Y = var_name_list$EEAs[i], 
                                X = exp, 
                                adjust_cfd = adjust_cfd, 
                                dat=df)
      plot_df[i,] <- rlts
    }
    colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
    rownames(plot_df) <- var_name_list$EEAs
    plot_df$EAAs <- EAAs_names
    plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
  
    plot_df$coefficient = as.numeric(plot_df$coefficient)
    plot_df$ci_lower = as.numeric(plot_df$ci_lower)
    plot_df$ci_upper = as.numeric(plot_df$ci_upper)
  
    lm_coef_ambient_plot <- make_forest_plot(plot_df, exp)
    
    plot_df_list[[exp]] <- plot_df
    lm_coef_ambient_plot_list[[exp]] = lm_coef_ambient_plot
  }
  return(list(plot_df_list, lm_coef_ambient_plot_list))
}
```

# 1. Description of study population

There are `r nrow(df)` visits with corresponding epigenetic ages available among `r length(unique(df$SbjctD))` female subjects. For these `r length(unique(df$SbjctD))` subjects, `r length(unique(df$SbjctD)) - length(dup_sbjctD)` have been visited once and `r length(dup_sbjctD)` have been visited twice. 

The following tables summarize all the information of the first visit of these `r length(unique(df$SbjctD))` subjects.

## Baseline characteristics (confounders)
```{r}
df %>% select(var_name_list$confounders) %>% mutate(ses = as.factor(ses)) %>%
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```


## Epigenetic ages
```{r}
df %>% select(var_name_list$clocks) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

## Epigenetic ages accelarations
```{r}
df %>% select(var_name_list$EEAs) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

## Fuel/stove type exposures
```{r}
df %>% select(var_name_list$fuel_exp, var_name_list$stove_exp) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

## 5MC exposures
```{r}
df %>% select(var_name_list$PCH_5MC) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

## Cluster-based exposures
### clusCUR6
Clusters based on model-based exposure estimates at or shortly before the visit
```{r}
df %>% select(var_name_list$cluster_exp$clusCUR6) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

### clusCHLD5
Clusters based on model-based exposure estimates accrued before age 18 
```{r}
df %>% select(var_name_list$cluster_exp$clusCHLD5) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

### clusCUM6
Clusters based on model-based lifetime exposure estimates 
```{r}
df %>% select(var_name_list$cluster_exp$clusCUM6) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

### clusMEAS6
Clusters based on pollutant measurements
```{r}
df %>% select(var_name_list$cluster_exp$clusMEAS6) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

### clusURI5
Clusters based on urinary biomarkers
```{r}
df %>% select(var_name_list$cluster_exp$clusURI5) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```



## Ambient exposures
```{r}
df %>% select(var_name_list$ambient_exp) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

## Urinary biomarkers
```{r}
df %>% select(var_name_list$urinary_exp) %>% 
  tbl_summary(statistic=list(all_continuous() ~ "{mean} ({sd})", 
                             all_categorical() ~ "{n} / {N} ({p}%)"),
              missing_text = "(Missing)")
```

# 2. Self-reported categorical fuel type
## a.	Primary analysis 
### i. Current fuel type
The numbers of observations with each current fuel type:
```{r}
table(df$curFuel)
```
#### GEE (Mix)
In this section, we perform the generalized estimating equations (GEE) to evaluate the associations between current fuel type and each Epigenetic Age Acceleration with the formula: 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 * I(\text{Smokeles}) + \beta_2 * I(\text{Wood_and_or_Plant})   \\
  & + \beta_3 * county + \beta_4 * BMI + \beta_5 * ses + \beta_6 * edu + \epsilon
\end{aligned}
$$


Results:
```{r}
preform_gee_curFuel <- function(Y, adjust_cfd = F) {
  X = "curFuel"
  tmp <- df[!is.na(df[,X]),]
  
  tmp$curFuel = factor(tmp$curFuel, levels = c("Smoky","Smokeles","Wood_and_or_Plant"))
  
  form <- ifelse(adjust_cfd, 
                 paste0(Y," ~ ", X, " + ", paste(var_name_list$confounders[-1], collapse  = " + ")),
                 paste0(Y,"~",X))
  form <- as.formula(form)
  
  geefit <- geeglm(formula = form, 
                data = tmp,
                id = SbjctD,
                family = gaussian,
                corstr = "exchangeable") 

  coefs <- as.data.frame(summary(geefit)$coefficients)[c(1,2,3), c(1,2,4)]
  colnames(coefs) <- c("coefficient", "std", "p_val")
  rownames(coefs)[rownames(coefs) == "(Intercept)"] = "Smoky (reference/intercept)"
  rownames(coefs)[rownames(coefs) == "curFuelSmokeles"] = "Smokeles"
  rownames(coefs)[rownames(coefs) == "curFuelWood_and_or_Plant"] = "Wood_and_or_Plant"  

  coefs <- coefs %>%
    mutate(ci_lower = coefficient - 1.96 * std,
           ci_upper = coefficient + 1.96 * std,
           sig_level = ifelse(p_val <= 0.001, "<= 0.001",
                        ifelse(p_val <= 0.01, "<= 0.01",
                               ifelse(p_val <= 0.05, "<= 0.05", "> 0.05")))) %>% 
    select(coefficient, std, ci_lower, ci_upper, p_val, sig_level)
  return(coefs)
}
```

```{r gee_curFuel_mix, fig.width=7, fig.height=3}
fueltype_smoky_mean <- data.frame(matrix(NA, ncol = 6))
fueltype_smokeless_mean <- data.frame(matrix(NA, ncol = 6))
fueltype_woodplant_mean <- data.frame(matrix(NA, ncol = 6))

for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee_curFuel(Y = var_name_list$EEAs[i], adjust_cfd = T)
    fueltype_smoky_mean[i,] <- rlts[1,]
    fueltype_smokeless_mean[i,] <- rlts[2,]
    fueltype_woodplant_mean[i,] <- rlts[3,]
}

colnames(fueltype_smoky_mean) = colnames(rlts)
colnames(fueltype_smokeless_mean) = colnames(rlts)
colnames(fueltype_woodplant_mean) = colnames(rlts)

rownames(fueltype_smoky_mean) = var_name_list$EEAs
rownames(fueltype_smokeless_mean) = var_name_list$EEAs
rownames(fueltype_woodplant_mean) = var_name_list$EEAs

fueltype_smoky_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
fueltype_smokeless_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
fueltype_woodplant_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

p1 <- make_forest_plot_lg(fueltype_smoky_mean, "Smoky (reference)")
p2 <- make_forest_plot_lg(fueltype_smokeless_mean, "Smokeless")
p3 <- make_forest_plot_lg(fueltype_woodplant_mean, "Wood_and_or_Plant")

plot_list <- list(p2, p3)

for(j in 1:2){
  plot_list[[j]] <- plot_list[[j]] + ylim(c(-7, 7))
}

grid.arrange(grobs = plot_list,  ncol = 2,
             top = textGrob(paste("Average EAA difference of each current Fuel Type\n compared to Smoky Coal"), 
                            gp=gpar(fontsize=15, fontface = 'bold')))

```

```{r}
coefs <- preform_gee_curFuel(Y = "AgeAccelGrim", adjust_cfd = T)
cat("The estimated average GrimAge EAA difference of current fuel type, smokeless/wood or plant compared to smoky coal:")
print(coefs)
```


#### GEE (no confounders)
In this section, we perform the generalized estimating equations (GEE) to evaluate the associations between current fuel type and each Epigenetic Age Acceleration with the formula: 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 * I(\text{Smokeles}) + \beta_2 * I(\text{Wood_and_or_Plant}) + \epsilon
\end{aligned}
$$


Results:
```{r gee_curFuel_simple, fig.width=7, fig.height=3}
fueltype_smoky_mean <- data.frame(matrix(NA, ncol = 6))
fueltype_smokeless_mean <- data.frame(matrix(NA, ncol = 6))
fueltype_woodplant_mean <- data.frame(matrix(NA, ncol = 6))

for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee_curFuel(Y = var_name_list$EEAs[i], adjust_cfd = F)
    fueltype_smoky_mean[i,] <- rlts[1,]
    fueltype_smokeless_mean[i,] <- rlts[2,]
    fueltype_woodplant_mean[i,] <- rlts[3,]
}

colnames(fueltype_smoky_mean) = colnames(rlts)
colnames(fueltype_smokeless_mean) = colnames(rlts)
colnames(fueltype_woodplant_mean) = colnames(rlts)

rownames(fueltype_smoky_mean) = var_name_list$EEAs
rownames(fueltype_smokeless_mean) = var_name_list$EEAs
rownames(fueltype_woodplant_mean) = var_name_list$EEAs

fueltype_smoky_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
fueltype_smokeless_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
fueltype_woodplant_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

p1 <- make_forest_plot_lg(fueltype_smoky_mean, "Smoky (reference)")
p2 <- make_forest_plot_lg(fueltype_smokeless_mean, "Smokeless")
p3 <- make_forest_plot_lg(fueltype_woodplant_mean, "Wood_and_or_Plant")

plot_list <- list(p2, p3)

for(j in 1:2){
  plot_list[[j]] <- plot_list[[j]] + ylim(c(-7, 7))
}

grid.arrange(grobs = plot_list,  ncol = 2,
             top = textGrob(paste("Average EAA difference of each current Fuel Type\n compared to Smoky Coal"), 
                            gp=gpar(fontsize=15, fontface = 'bold')))

```

```{r}
coefs <- preform_gee_curFuel(Y = "AgeAccelGrim", adjust_cfd = T)
cat("The estimated average GrimAge EAA difference of current fuel type, smokeless/wood or plant compared to smoky coal:")
print(coefs)
```




## b.	Secondary  analyses 
Additional analyses to present in tandem with the primary analysis findings given that cumulative and long-term exposure may impact methylation

### i. Cumulative fuel type
Average EAA difference of cumulative fuel type, mixed fuel compared to smoky coal-significant association for GrimAge 

The numbers of observations with each current fuel type:
```{r}
table(df$cumFuel)
```


#### GEE (Mix)
In this section, we perform the generalized estimating equations (GEE) to evaluate the associations between cumulative fuel type and each Epigenetic Age Acceleration with the formula: 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 * I(\text{Mix}) \\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu + \epsilon
\end{aligned}
$$


Results:
```{r}
preform_gee_cumFuel <- function(Y, adjust_cfd = F) {
  X = "cumFuel"
  tmp <- df[!is.na(df[,X]),]
  
  tmp$cumFuel = factor(tmp$cumFuel, levels = c("Smoky","Mix"))
  
  form <- ifelse(adjust_cfd, 
                 paste0(Y," ~ ", X, " + ", paste(var_name_list$confounders[-1], collapse  = " + ")),
                 paste0(Y,"~",X))
  form <- as.formula(form)
  
  geefit <- geeglm(formula = form, 
                data = tmp,
                id = SbjctD,
                family = gaussian,
                corstr = "exchangeable") 

  coefs <- as.data.frame(summary(geefit)$coefficients)[c(1,2), c(1,2,4)]
  colnames(coefs) <- c("coefficient", "std", "p_val")
  rownames(coefs)[rownames(coefs) == "(Intercept)"] = "Smoky (reference/intercept)"
  rownames(coefs)[rownames(coefs) == "cumFuelMix"] = "Mix"
  coefs <- coefs %>%
    mutate(ci_lower = coefficient - 1.96 * std,
           ci_upper = coefficient + 1.96 * std,
           sig_level = ifelse(p_val <= 0.001, "<= 0.001",
                        ifelse(p_val <= 0.01, "<= 0.01",
                               ifelse(p_val <= 0.05, "<= 0.05", "> 0.05")))) %>% 
    select(coefficient, std, ci_lower, ci_upper, p_val, sig_level)
  return(coefs)
}
```

```{r gee_cumFuel_mix}
fueltype_smoky_mean <- data.frame(matrix(NA, ncol = 7))
fueltype_mix_mean <- data.frame(matrix(NA, ncol = 7))

for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee_cumFuel(Y = var_name_list$EEAs[i], adjust_cfd = T)
    fueltype_smoky_mean[i,] <- rlts[1,]
    fueltype_mix_mean[i,] <- rlts[2,]
}

colnames(fueltype_smoky_mean) = colnames(rlts)
colnames(fueltype_mix_mean) = colnames(rlts)

rownames(fueltype_smoky_mean) = var_name_list$EEAs
rownames(fueltype_mix_mean) = var_name_list$EEAs

fueltype_smoky_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
fueltype_mix_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

p1 <- make_forest_plot_lg(fueltype_smoky_mean, "Smoky (reference)")
p2 <- make_forest_plot_lg(fueltype_mix_mean, "Mix")

# plot_list <- list(p1, p2)
# 
# for(j in 1){
#   plot_list[[j]] <- plot_list[[j]] + ylim(c(-10, 10))
# }

p2 <- p2 + labs(title = "Average EAA difference of cumulative Fuel Type Mix\n compared to Smoky Coal",
                plot.title = element_text(size=12))
p2
# grid.arrange(p2,  ncol = 1,
#              top = textGrob(paste("Average EAA difference of cumulative Fuel Type Mix\n compared to Smoky Coal"), gp=gpar(fontsize=12, fontface = 'bold')))
```

```{r}
coefs <- preform_gee_cumFuel(Y = "AgeAccelGrim", adjust_cfd = T)
cat("The estimated average GrimAge EAA difference of cumulative fuel type, mix compared to smoky coal:")
print(coefs)
```


### ii. Childhood fuel type
Average EAA difference for childhood fuel type, smokeless/wood/mixed fuel compared to smoky

The numbers of observations with each childhood fuel type:
```{r}
table(df$brthFuel)
```

### GEE (Mix) 
In this section, we perform the generalized estimating equations (GEE) to evaluate the association between current fuel type and the Grim Epigenetic Age Acceleration with the formula: 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 * I(\text{Wood}) + \beta_2 * I(\text{Smokeles}) + \beta_3 * I(\text{Mix})  \\
  & + \beta_4 * county + \beta_5 * BMI + \beta_6 * ses + \beta_7 * edu + \epsilon
\end{aligned}
$$

Results:
```{r}
preform_gee_brthFuel <- function(Y, adjust_cfd = F) {
  X = "brthFuel"
  tmp <- df[!is.na(df[,X]),]
  
  tmp$brthFuel = factor(tmp$brthFuel, levels = c("Smoky","Wood","Smokeles", "Mix"))
  
  form <- ifelse(adjust_cfd, 
                 paste0(Y," ~ ", X, " + ", paste(var_name_list$confounders[-1], collapse  = " + ")),
                 paste0(Y,"~",X))
  form <- as.formula(form)
  
  geefit <- geeglm(formula = form, 
                data = tmp,
                id = SbjctD,
                family = gaussian,
                corstr = "exchangeable") 

  coefs <- as.data.frame(summary(geefit)$coefficients)[c(1,2,3,4), c(1,2,4)]
  colnames(coefs) <- c("coefficient", "std", "p_val")
  rownames(coefs)[rownames(coefs) == "(Intercept)"] = "Smoky (reference/intercept)"
  rownames(coefs)[rownames(coefs) == "brthFuelWood"] = "Wood"
  rownames(coefs)[rownames(coefs) == "brthFuelSmokeles"] = "Smokeles"
  rownames(coefs)[rownames(coefs) == "brthFuelMix"] = "Mix"

  coefs <- coefs %>%
    mutate(ci_lower = coefficient - 1.96 * std,
           ci_upper = coefficient + 1.96 * std,
           sig_level = ifelse(p_val <= 0.001, "<= 0.001",
                        ifelse(p_val <= 0.01, "<= 0.01",
                               ifelse(p_val <= 0.05, "<= 0.05", "> 0.05")))) %>% 
    select(coefficient, std, ci_lower, ci_upper, p_val, sig_level)
  return(coefs)
}
```

```{r gee_brthFuel_mix, fig.width=7, fig.height=5}

fueltype_smoky_mean <- data.frame(matrix(NA, ncol = 6))
fueltype_wood_mean <- data.frame(matrix(NA, ncol = 6))
fueltype_smokeless_mean <- data.frame(matrix(NA, ncol = 6))
fueltype_mix_mean <- data.frame(matrix(NA, ncol = 6))

for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee_brthFuel(Y = var_name_list$EEAs[i], adjust_cfd = T)
    fueltype_smoky_mean[i,] <- rlts[1,]
    fueltype_wood_mean[i,] <- rlts[2,]
    fueltype_smokeless_mean[i,] <- rlts[3,]
    fueltype_mix_mean[i,] <- rlts[4,]
}

colnames(fueltype_smoky_mean) = colnames(rlts)
colnames(fueltype_wood_mean) = colnames(rlts)
colnames(fueltype_smokeless_mean) = colnames(rlts)
colnames(fueltype_mix_mean) = colnames(rlts)

rownames(fueltype_smoky_mean) = EAAs_vars
rownames(fueltype_wood_mean) = EAAs_vars
rownames(fueltype_smokeless_mean) = EAAs_vars
rownames(fueltype_mix_mean) = EAAs_vars

fueltype_smoky_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
fueltype_wood_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
fueltype_smokeless_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))
fueltype_mix_mean$EAAs = factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

p1 <- make_forest_plot_lg(fueltype_smoky_mean, "Smoky (reference)")
p2 <- make_forest_plot_lg(fueltype_wood_mean, "Wood")
p3 <- make_forest_plot_lg(fueltype_smokeless_mean, "Smokeless")
p4 <- make_forest_plot_lg(fueltype_mix_mean, "Mix")

plot_list <- list(p2, p3, p4)

for(j in 1:3){
  plot_list[[j]] <- plot_list[[j]] + ylim(c(-7.5, 7.5))
}

grid.arrange(grobs = plot_list,  ncol = 2, 
             layout_matrix = rbind(c(1,2),
                                   c(3,NA)),
             top = textGrob(paste("Average EAA difference of each Childhood Fuel Type\n compared to Smoky Coal"), 
                            gp=gpar(fontsize=15, fontface = 'bold')))

```

```{r}
coefs <- preform_gee_brthFuel(Y = "AgeAccelGrim", adjust_cfd = T)
cat("The estimated average GrimAge EAA difference of childhood fuel type, smokeless/wood/mixed compared to smoky coal:")
print(coefs)
```



# 3. Cluster-based analyses
## Primary analysis
### Each cluster within current pollutant exposures 
CUR6_BC_PAH6 – Black carbon (BC) and 6 PAHs   
CUR6_PAH31 – a large cluster of 31 PAHs  
CUR6_NkF – NkF only  
CUR6_PM_RET – Particulate matter (PM) and retene  
CUR6_NO2 – NO2 only  
CUR6_SO2 – SO2 only    


Summary the exposure estimates:
```{r}
# summary(df %>% select(var_name_list$cluster_exp$clusCUR6))
df %>% 
  select(var_name_list$cluster_exp$clusCUR6, curFuel) %>%
  tbl_summary(by = curFuel,
              # type = all_continuous() ~ "continuous2",
              statistic=list(all_continuous() ~ "{median} ({p25}, {p75})"),
              digits = list(  ~ c(2, 1, 1)),
              missing_text = "(Missing)") %>% 
  add_overall()
```

### GEE (Mix)
In this section, we perform the generalized estimating equations (GEE) to evaluate the associations between each cluster within current pollutant exposures and each Epigenetic Age Acceleration with the formula: 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 * X   \\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations and X is one of the cluster estimates. 


Results:
```{r gee_clusCUR6_mix, fig.width=8, fig.height=7}
rlts_gee <- perform_group_gee(X = var_name_list$cluster_exp$clusCUR6, adjust_cfd = T)

for(j in 1:length(rlts_gee[[2]])){
  rlts_gee[[2]][[j]] <- rlts_gee[[2]][[j]] + ylim(c(-2.2, 2.2))
}

grid.arrange(grobs = rlts_gee[[2]],  ncol = 2,
             top = textGrob(paste("Estimated Effect of each exposure prototype \n based on current pollutant exposures"), gp=gpar(fontsize=15, fontface = 'bold')))
```
```{r}
cat("The estimated effects:")
for(j in 1:length(rlts_gee[[1]])){
  rlts_gee[[1]][[j]] <- rlts_gee[[1]][[j]][, -7]
}
print(rlts_gee[[1]])
```

## b.	Secondary analyses 
### i. Each cluster within cumulative lifetime pollutant exposures
Effect of each cluster for cumulative lifetime pollutant exposures and EAA-significant findings for PAH36 (PhenoAge, skin & blood) and NkF (GrimAge)


CUM6_BC_NO2_PM – a cluster of BC, NO2, and PM  
CUM6_PAH36 – a large cluster of 36 PAHs  
CUM6_DlP – DlP only  
CUM6_NkF – NkF only  
CUM6_RET – retene only  
CUM6_SO2 – SO2 only   


Summary the exposure estimates:
```{r}
# summary(df %>% select(var_name_list$cluster_exp$clusCUR6))
df %>% 
  select(var_name_list$cluster_exp$clusCUM6, curFuel) %>%
  tbl_summary(by = curFuel,
              # type = all_continuous() ~ "continuous2",
              statistic=list(all_continuous() ~ "{median} ({p25}, {p75})"),
              digits = list(  ~ c(2, 1, 1)),
              missing_text = "(Missing)") %>% 
  add_overall()
```

In this section, we perform the generalized estimating equations (GEE) to evaluate the associations between each cluster within current pollutant exposures and each Epigenetic Age Acceleration with the formula: 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 * X   \\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations and X is one of the cluster estimates. 


Results:
```{r gee_clusCUM6_mix, fig.width=8, fig.height=7}
rlts_gee <- perform_group_gee(X = var_name_list$cluster_exp$clusCUM6, adjust_cfd = T)

for(j in 1:length(rlts_gee[[2]])){
  rlts_gee[[2]][[j]] <- rlts_gee[[2]][[j]] + ylim(c(-2.2, 2.2))
}

grid.arrange(grobs = rlts_gee[[2]],  ncol = 2,
             top = textGrob(paste("Estimated Effect of each exposure prototype \n based on cumulative pollutant exposures"), gp=gpar(fontsize=15, fontface = 'bold')))
```
```{r}
cat("The estimated effect:")
for(j in 1:length(rlts_gee[[1]])){
  rlts_gee[[1]][[j]] <- rlts_gee[[1]][[j]][, -7]
}
print(rlts_gee[[1]])
```



### ii. Each cluster within childhood pollutant exposures 
Effect of each cluster with childhood pollutant exposures and EAA-significant LRT P value for GrimAge but no findings for linear regression

Summary the exposure estimates:
```{r}
# summary(df %>% select(var_name_list$cluster_exp$clusCUR6))
df %>% 
  select(var_name_list$cluster_exp$clusCHLD5, curFuel) %>%
  tbl_summary(by = curFuel,
              # type = all_continuous() ~ "continuous2",
              statistic=list(all_continuous() ~ "{median} ({p25}, {p75})"),
              digits = list(  ~ c(2, 1, 1)),
              missing_text = "(Missing)") %>% 
  add_overall()
```

In this section, we perform the generalized estimating equations (GEE) to evaluate the associations between each cluster within current pollutant exposures and each Epigenetic Age Acceleration with the formula: 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 * X   \\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations and X is one of the cluster estimates. 


Results:
```{r gee_clusCHLD5_mix, fig.width=8, fig.height=7}
rlts_gee <- perform_group_gee(X = var_name_list$cluster_exp$clusCHLD5, adjust_cfd = T)

for(j in 1:length(rlts_gee[[2]])){
  rlts_gee[[2]][[j]] <- rlts_gee[[2]][[j]] + ylim(c(-3, 3))
}

grid.arrange(grobs = rlts_gee[[2]],  ncol = 2,
             top = textGrob(paste("Estimated Effect of each exposure prototype \n based on childhood pollutant exposures"), gp=gpar(fontsize=15, fontface = 'bold')))
```
```{r}
cat("The estimated effect:")
for(j in 1:length(rlts_gee[[1]])){
  rlts_gee[[1]][[j]] <- rlts_gee[[1]][[j]][, -7]
}
print(rlts_gee[[1]][1:2])
```




# 4. 5 methylchrysene (5MC) 


## Summary the exposure estimates:
```{r}
df %>% 
  select(var_name_list$PCH_5MC, curFuel) %>% 
  tbl_summary(by = curFuel,
              # type = all_continuous() ~ "continuous2",
              statistic=list(all_continuous() ~ "{median} ({p25}, {p75})"),
              digits = list(  ~ c(2, 1, 1)),
              missing_text = "(Missing)") %>% 
  add_overall()
```

```{r}
print("Pearson pair-wise correlation:")
cor(df %>%  select(var_name_list$PCH_5MC) %>% na.omit(), 
    method = "pearson")
```
```{r}
print("Spearman pair-wise correlation:")
cor(df %>%  select(var_name_list$PCH_5MC) %>% na.omit(), 
    method = "spearman")
```


## Primary analyses
Effect of modeled current exposure to 5MC and EAA

### Generalized estimating equations (GEE) 
#### No confounders
In the following section, we performed generalized estimating equations (GEE) with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *cur\_5mc + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and current $5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in current 5MC while holding other variables constant".
```{r gee_coef_cur_5MC_simple, fig.width=4, fig.height=3, warning=FALSE, message=FALSE}
  
  lm_coef_ambient <- c()
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee(Y = var_name_list$EEAs[i], 
                              X = "cur_5mc", 
                              adjust_cfd = F, 
                              dat=df)
    lm_coef_ambient[i] <- rlts[1]
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Current exposure to 5MC")

# lm_coef_ambient[9,1] <- c("[P<0.001: ***; P<0.01: **; P<0.05: *]")
# colnames(lm_coef_ambient) <- "PCH_5MC"
# rownames(lm_coef_ambient) <- c(EAAs_names, "")
print(" Result data: ")
print(plot_df)
lm_coef_ambient_plot
```


#### Mixed model
In the following section, we performed generalized estimating equations (GEE) with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *cur\_5mc\\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu  + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and current $5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in current 5MC while holding other variables constant".
```{r gee_coef_cur_5MC, fig.width=4, fig.height=3, warning=FALSE, message=FALSE}
  
  lm_coef_ambient <- c()
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee(Y = var_name_list$EEAs[i], 
                              X = "cur_5mc", 
                              adjust_cfd = T, 
                              dat=df)
    lm_coef_ambient[i] <- rlts[1]
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Current exposure to 5MC")

# lm_coef_ambient[9,1] <- c("[P<0.001: ***; P<0.01: **; P<0.05: *]")
# colnames(lm_coef_ambient) <- "PCH_5MC"
# rownames(lm_coef_ambient) <- c(EAAs_names, "")
print(" Result data: ")
print(plot_df)
lm_coef_ambient_plot
```



### Linear Regression (mix model)
In the following section, we performed linear regression with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *cur\_5mc\\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu  + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and $5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in current 5MC while holding other variables constant".

#### Including all visits
```{r linear_coef_cur_5MC, fig.width=4, fig.height=3}
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_regression(Y = var_name_list$EEAs[i], 
                              X = "cur_5mc", 
                              adjust_cfd = T, 
                              dat=df)
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Current exposure to 5MC")

  print(" Result data: ")
  print(plot_df)
  lm_coef_ambient_plot
```

#### Only including one of the two visits 
```{r linear_coef_cur_5MC_subs, fig.width=3, fig.height=4}
lm_coef_ambient <- data.frame()
lm_coef_ambient_plots <- list()

for(j in 1:2){
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_regression(Y = var_name_list$EEAs[i], 
                              X = "cur_5mc", 
                              adjust_cfd = T, 
                              dat=df_subs[[j]])
    lm_coef_ambient[i,j] <- rlts[1]
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  pj <- make_forest_plot(plot_df, "Current exposure to 5MC")
  lm_coef_ambient_plots[[j]] = pj
}


lm_coef_ambient[9,1] <- c("[P<0.001: ***; P<0.01: **; P<0.05: *]")
colnames(lm_coef_ambient) <- "PCH_5MC"
rownames(lm_coef_ambient) <- c(EAAs_names, "")
```

Including all of the single visits and the 1st visit for these `r length(dup_sbjctD)` subjects who have been visited twice. 
```{r linear_coef_cur_5MC_sub_1, fig.height=3, fig.width=4}
lm_coef_ambient_plots[[1]]
```

Including all of the single visits and the 2nd visit for these `r length(dup_sbjctD)` subjects who have been visited twice. 
```{r linear_coef_cur_5MC_sub_2, fig.height=3, fig.width=4}
lm_coef_ambient_plots[[2]]
```



## Secondary analyses
### i. Effect of modeled cumulative exposure to 5MC and EAA
#### Generalized estimating equations (GEE) 
In the following section, we performed generalized estimating equations (GEE) with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *cum\_5mc\\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and cumulative $5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in cumulative 5MC while holding other variables constant".
```{r gee_coef_cum_5MC, fig.width=4, fig.height=3, warning=FALSE, message=FALSE}
  
  lm_coef_ambient <- c()
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee(Y = var_name_list$EEAs[i], 
                              X = "cum_5mc", 
                              adjust_cfd = T, 
                              dat=df)
    lm_coef_ambient[i] <- rlts[1]
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Cumulative exposure to 5MC")

# lm_coef_ambient[9,1] <- c("[P<0.001: ***; P<0.01: **; P<0.05: *]")
# colnames(lm_coef_ambient) <- "PCH_5MC"
# rownames(lm_coef_ambient) <- c(EAAs_names, "")
print(" Result data: ")
print(plot_df)
lm_coef_ambient_plot
```


#### Linear Regression (mix model)
In the following section, we performed linear regression with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *cum\_5mc\\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu  + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and $5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in cumulative 5MC while holding other variables constant".

##### Including all visits
```{r linear_coef_cum_5MC, fig.width=4, fig.height=3}
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_regression(Y = var_name_list$EEAs[i], 
                              X = "cum_5mc", 
                              adjust_cfd = T, 
                              dat=df)
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Cumulative exposure to 5MC")

  print(" Result data: ")
  print(plot_df)
  lm_coef_ambient_plot
```


### ii. Effect of modeled birth home/childhood exposure to 5MC
#### Generalized estimating equations (GEE) 
In the following section, we performed generalized estimating equations (GEE) with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *birth\_5mc\\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu  + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and birth home/childhood  $5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in birth home/childhood 5MC while holding other variables constant".
```{r gee_coef_birth_5MC, fig.width=4, fig.height=3, warning=FALSE, message=FALSE}
  lm_coef_ambient <- c()
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee(Y = var_name_list$EEAs[i], 
                              X = "bir_5mc", 
                              adjust_cfd = T, 
                              dat=df)
    lm_coef_ambient[i] <- rlts[1]
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Birth home/childhood exposure to 5MC")

# lm_coef_ambient[9,1] <- c("[P<0.001: ***; P<0.01: **; P<0.05: *]")
# colnames(lm_coef_ambient) <- "PCH_5MC"
# rownames(lm_coef_ambient) <- c(EAAs_names, "")
print(" Result data: ")
print(plot_df)
lm_coef_ambient_plot
```



#### Linear Regression (mix model)
In the following section, we performed linear regression with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *birth\_5mc\\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu  + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and birth home/childhood $5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in birth home/childhood 5MC while holding other variables constant".

##### Including all visits
```{r linear_coef_birth_5MC, fig.width=4, fig.height=3}
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_regression(Y = var_name_list$EEAs[i], 
                              X = "bir_5mc", 
                              adjust_cfd = T, 
                              dat=df)
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Birth home/childhood exposure to 5MC")

  print(" Result data: ")
  print(plot_df)
  lm_coef_ambient_plot
```


### iii.	Effect of measured current exposure to 5MC
#### Generalized estimating equations (GEE) 
In the following section, we performed generalized estimating equations (GEE) with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *cur\_measured\_5mc\\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu  + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and measured current exposure to  $5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in measured current exposure to 5MC while holding other variables constant".
```{r gee_coef_cum_measured_5MC, fig.width=4, fig.height=3, warning=FALSE, message=FALSE}
  
  lm_coef_ambient <- c()
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_gee(Y = var_name_list$EEAs[i], 
                              X = "cur_5mc_measured", 
                              adjust_cfd = T, 
                              dat=df)
    lm_coef_ambient[i] <- rlts[1]
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Measured current exposure to 5MC")

# lm_coef_ambient[9,1] <- c("[P<0.001: ***; P<0.01: **; P<0.05: *]")
# colnames(lm_coef_ambient) <- "PCH_5MC"
# rownames(lm_coef_ambient) <- c(EAAs_names, "")
print(" Result data: ")
print(plot_df)
lm_coef_ambient_plot
```



#### Linear Regression (mix model)
In the following section, we performed linear regression with equation 
$$
\begin{aligned}
  Y = & \beta_0 + \beta_1 *cum\_5mc\\
  & + \beta_2 * county + \beta_3 * BMI + \beta_4 * ses + \beta_5 * edu  + \epsilon
\end{aligned}
$$
where $Y$ is one of the epigenetic age accelerations.

The estimations of $\beta_1$ with given $Y$ and measured current exposure to 5MC$ are shown below, which can be interpreted as "the mean of Y changes given a one-unit increase in measured current exposure to  5MC while holding other variables constant".

##### Including all visits
```{r linear_coef_cur_measured_5MC, fig.width=4, fig.height=3}
  plot_df <- data.frame(matrix(NA, ncol = 6))
  for(i in 1:length(var_name_list$EEAs)){
    rlts = preform_regression(Y = var_name_list$EEAs[i], 
                              X = "cur_5mc_measured", 
                              adjust_cfd = T, 
                              dat=df)
    plot_df[i,] <- rlts
  }
  colnames(plot_df) <- c("coefficient", "std", "ci_lower", "ci_upper", "p_val", "sig_level")
  rownames(plot_df) <- var_name_list$EEAs
  plot_df$EAAs <- EAAs_names
  plot_df$EAAs <- factor(EAAs_names, levels=rev(c(EAAs_names[EAAs_names != "DNAmTL"], "DNAmTL")))

  plot_df$coefficient = as.numeric(plot_df$coefficient)
  plot_df$ci_lower = as.numeric(plot_df$ci_lower)
  plot_df$ci_upper = as.numeric(plot_df$ci_upper)

  lm_coef_ambient_plot <- make_forest_plot(plot_df, "Measured current exposure to 5MC")

  print(" Result data: ")
  print(plot_df)
  lm_coef_ambient_plot
```

# 5. Additional secondary analyses
